Add thick / thin library variants
The thick (default) build contains AIDL and Parcelable classes that are
included, but not exposed in the GCM library. These classes are
currently necessary in order to communicate with the scheduling engine
inside Google Play services. Instead of requiring a full dependency on
the GCM library, the thick variant includes the two required classes.

This unfortunately introduces a duplicate class error for any user that
includes both the GCM library and this library (#3). To solve this issue
for those users, the thin variant doesn't include the copied and hidden
classes. Instead, it relies on an explicit dependency on the GCM library
to provide those classes at runtime. Users who are required to have a
dependency on GCM ("com.google.android.gms:play-services-gcm") should
use this variant.