Handle errors when sending task results back
Most of the receiver code that communicates back with Google Play
services uses the sendResultsSafely method to correctly guard against
the possibility of a RemoteException occurring. Unfortunately the
routine that returns the result at the end of execution did not. This
can cause a DeadObjectException if Google Play services dies before the
task result is sent.

Update the onJobFinished callback to use sendResultsSafely instead of
potentially crashing the client app.

Fixes #26.