Call stopSelf after all jobs finish
The jobFinished method wasn't calling stopSelf, which led to the
GooglePlayReceiver (a Service) running until the system decided to stop
it. This adds a call to stopSelf (which is safe to do even if the
service is being bound to).

Also added some tests for the onStartCommand implementation and
refactored the some common elements out of the
GooglePlayCallbackExtractorTest.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=158755156