Fix JobDispatcher's pre-L callback extraction
Something about the way Bundle keys are written to Parcels got changed
during the run-up to L. Before Lollipop, keys had to be read using
Parcel#readValue(ClassLoader). On Lollipop and after, keys have to be
read using Parcel#readString().

This patch fixes that, and also adds a set of android_test rules to
validate the end-to-end experience on (almost) all supported API levels.

This is causing issues for 1st party apps ([] but the
underlying code change that caused the issue hasn't been exported to the
public repo yet. A followup CL will be required to fix the Gradle build
prior to exporting to GitHub.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=159766194