Convert JobService.LocalBinder to an AIDL interface
This fixes support for JobService in non-default processes. Previously we were
using the JobService.LocalBinder class, which relies on being run in the same
process. This patch migrates to an AIDL interface, which standardizes the
interface and allows the GooglePlayReceiver to communicate with a JobService in
a separate process.

Also fixes an issue in the JobCoder which caused internal keys to be
written to the same extras Bundle that clients used. Now extras are
always written and extracted. The EndToEnd test has been adjusted to
validate that the extras received in onStartJob match the ones delivered
in the initial schedule request.