Add ConstraintChecker to Firebase Job Dispatcher
This adds a new "ConstraintChecker" class to the library. This class is
responsible for validating that a job execution constraints are
satisfied prior to starting jobs.

This is useful for checks that the Google Play services scheduling
engine is unable to definitively determine such as app network
accessibility when Data Saver Mode is enabled.

This change just adds the ConstraintChecker. A future change will update
the ExecutionDelegator to make use of the ConstraintChecker.

The initial version of the ConstraintChecker only looks for satisfaction
of network constraints. To do this, we need the ACCESS_NETWORK_STATE
permission which is a normal permission that Android can automatically
grant to apps at install time[1].

[1]: https://developer.android.com/guide/topics/permissions/overview.html#normal-dangerous

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=187928253