Use ContraintChecker to guard job execution
We use the recently added ConstraintChecker to validate that a job's
constraints are satisfied prior to starting the job. If constraints are
fulfilled, execution proceeds directly as before.

If the constraints are unsatisfied, the job execution result is marked
as RESULT_FAIL_RETRY. This is so that the job can be retried later on
for execution if eligible.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=188258796