Catch SecurityExceptions when binding within app
Under certain circumstances specific builds of Android seem to reject
calls to bindService with a SecurityException, even when targeting a
Service that exists within the same app. This patch changes the
ExecutionDelegator to catch those exceptions instead of crashing the app
(and causing a timeout on the Google Play services side). It also
ensures that binding fails will cause a RETRY result to be sent back to
the caller, which should help reduce the overall number of timeouts when
using the FJD.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=195499014